requirements-python2.txt
-e .
pycodestyle
pylint
pytest
pytest-click<1.0
pytest-cov
mock
